VET057-76 Added conditional rendering to Animal Page

+++ src/main/java/com/vetapp/application/prescription/PrescriptionRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PrescriptionRepository extends JpaRepository<Prescription, Long>{
+public interface PrescriptionRepository extends JpaRepository<Prescription, Integer>{
 
 
     List<Prescription> findByanimalid(int animalid);
